Increase kValueSize for CHERI architectures.
On 64-bit architectures the Value type stores a single pointer (with
tagging in the lower bits). To accomodate this kValueSize is set to a
value of 8. On 32-bit architectures a single pointer is stored with the
lower empty 32-bits used for tagging. On CHERI architectures the size of
kValueSize must be twice the native 64-bit archictectures integer pointer
size.